% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\name{Calculate_Kernel}
\alias{Calculate_Kernel}
\title{Calculate the kernel matrix.}
\usage{
Calculate_Kernel(
  object,
  approximation = FALSE,
  bandwidth.set.by.user = NULL,
  bandwidthtype = "Silverman",
  sparseKernel = FALSE,
  sparseKernel_tol = 1e-10,
  sparseKernel_ncore = 1
)
}
\arguments{
\item{object}{The Celina object.}

\item{approximation}{Option to use the default 11 kernels combined results or 1 kernel approximated results
True is to proceed with approximated algorithm}

\item{bandwidth.set.by.user}{User could select their own bandwidth (a numeric value) 
if the recommended bandwidth doesn't work in their dataset.}

\item{bandwidthtype}{The type of bandwidth to be used in Gaussian kernel, 
"SJ" for Sheather & Jones (1991) method (usually used in small size datasets), 
"Silverman" for Silverman's ‘rule of thumb’ method (1986) (usually used in large size datasets).}

\item{sparseKernel}{Select "TURE" if the user wants to use a sparse kernel matrix or "FALSE" if not. 
It is recommended to choose sparseKernel = "TRUE" when sample size is large.}

\item{sparseKernel_tol}{When sparseKernel = TRUE, 
the cut-off value when building sparse kernel matrix, 
any element in the kernel matrix greater than sparseKernel_tol will be kept, 
otherwise will be set to 0 to save memory.}

\item{sparseKernel_ncore}{When sparseKernel=TRUE, 
the number of CPU cores to build sparse kernel matrix.}
}
\value{
Returns Celina object.
}
\description{
Calculate the kernel matrix.
}
