% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\docType{class}
\name{Celina-class}
\alias{Celina-class}
\title{Each Celina object has a number of slots which store information. Key slots to access
are listed below.}
\description{
Each Celina object has a number of slots which store information. Key slots to access
are listed below.
}
\section{Slots}{

\describe{
\item{\code{celltype_mat}}{The raw celltype indicator or proportion matrix. 
Rows are celltypes, columns are spots/cells.}

\item{\code{gene_expression_mat}}{Original expression matrix. By default we use 
normalizeCounts normalization in Scater R package.}

\item{\code{genes_list}}{Different tested gene lists for cell types.}

\item{\code{project}}{Name of the project (for record keeping).}

\item{\code{covariates}}{The covariates in experiments (if any covariates is included).}

\item{\code{location}}{Cell/spot spatial coordinates to compute the kernel matrix.}

\item{\code{kernelmat}}{The kernel matrix for spatial relationship between locations.}

\item{\code{kernelmat_approx_U}}{The calculated approximated kernel matrix for spatial relationship between locations.}

\item{\code{bandwidth}}{The bandwidth in Gaussian kernel, users can also specify their preferred bandwidth.}

\item{\code{result}}{The summary of cell type specific SE genes results for different cell types.}

\item{\code{approximation}}{logical value to indicate which type of algorithms to be used in the analysis}
}}

