\name{CELINA-package}
\alias{CELINA-package}
\alias{CELINA}
\docType{package}
\title{\packageTitle{CELINA}}
\description{\packageDescription{CELINA}}
\section{Package Content}{\packageIndices{CELINA}}
\author{\packageAuthor{CELINA}}
\section{Maintainer}{\packageMaintainer{CELINA}}
\keyword{package}
