% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\name{Celina_kernel}
\alias{Celina_kernel}
\title{Construct the kernel matrix}
\usage{
Celina_kernel(location)
}
\arguments{
\item{location}{Cell/spot spatial coordinates to compute the kernel matrix.}
}
\value{
A list with 11 pre-calculated kernels
}
\description{
This function construct the kernel matrices in a list.
}
