% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\name{Celina_kernel_davies}
\alias{Celina_kernel_davies}
\title{Construct the kernel matrix used by davies p-value calculation}
\usage{
Celina_kernel_davies(kernels_list)
}
\arguments{
\item{kernels_list}{Pre-computed kernel matrices.}
}
\value{
A list with 11 new kernels
}
\description{
This function construct the kernel matrices to be used in the davies p-value calculation in a list.
}
