% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\name{Celina_test_davies}
\alias{Celina_test_davies}
\title{Testing multiple kernel matrices}
\usage{
Celina_test_davies(model1, kernel_mat = NULL, check_positive = TRUE)
}
\arguments{
\item{model1}{Celina fitted null models with parameters}

\item{kernel_mat}{A list to store the pre-defined kernel matrix,}

\item{check_positive}{Check the kernel matrix is positive or not}
}
\description{
Testing multiple kernel matrices
}
