% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\name{Celina_test_each_kernel}
\alias{Celina_test_each_kernel}
\title{Testing one kernel matrix to identify spatial pattern}
\usage{
Celina_test_each_kernel(
  model1,
  kernel_mat_each,
  precal_list = NULL,
  check_positive = FALSE
)
}
\arguments{
\item{model1}{Celina fitted null models with parameters}

\item{kernel_mat_each}{The single kernel matrix to be tested}

\item{precal_list}{Precalculated identities list to save computational time}

\item{check_positive}{Check the kernel matrix is positive or not}
}
\description{
Testing one kernel matrix to identify spatial pattern
}
