% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\name{Create_Celina_Object}
\alias{Create_Celina_Object}
\title{Create the Celina object.}
\usage{
Create_Celina_Object(
  celltype_mat = NULL,
  gene_expression_mat = NULL,
  location = NULL,
  covariates = NULL,
  project = "Celina"
)
}
\arguments{
\item{celltype_mat}{The raw celltype indicator or proportion matrix. Rows are celltypes, columns are spots/cells.}

\item{gene_expression_mat}{Original expression matrix. By default we use 
normalizeCounts normalization in Scater R package.}

\item{location}{Spatial location matrix (matrix), the dimension is n x d, n is the number of locations, d is dimensin of spatial coordinates, 
e.g. d = 2 for locations on 2D space. The rownames of locations and the colnames of count matrix should be matched.}

\item{covariates}{The covariates in experiments (matrix, if any covariates included), n x q, 
n is the number of locations, q is the number of covariates. 
The rownames of covariates and the rownames of locations should be matched.}

\item{project}{Name of the project (User defined).}
}
\value{
Returns Celina object.
}
\description{
Create the Celina object.
}
