% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\name{get_cell_type_gene_list}
\alias{get_cell_type_gene_list}
\title{Obtain a list of genes to test different cell types}
\usage{
get_cell_type_gene_list(
  spatial_count,
  cell_types_to_test,
  celltype_proportion,
  scRNA_count,
  sc_cell_type_labels,
  threshold
)
}
\arguments{
\item{spatial_count}{a gene expression matrix of spatial transcriptomics data with gene names as rownames and cells as columns}

\item{cell_types_to_test}{a vector of cell types to be used for testing}

\item{celltype_proportion}{a matrix of cell type proportions for each location in \code{spatial_count}, with locations as rows and cell types as columns}

\item{scRNA_count}{a gene expression matrix of reference scRNA-seq data with gene names as rownames and cells as columns}

\item{sc_cell_type_labels}{a vector of cell type labels for each cell in \code{scRNA_count}}

\item{threshold}{a number for filtering the genes}
}
\value{
cell_type_means, a data_frame of genes by cell types for mean normalized expression
}
\description{
Obtain a list of genes to test different cell types
}
