% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\name{preprocess_input}
\alias{preprocess_input}
\title{Preprocess the gene expression matrix - normalize the counts and filter genes}
\usage{
preprocess_input(
  object,
  cell_types_to_test,
  scRNA_count,
  sc_cell_type_labels,
  threshold = 5e-05
)
}
\arguments{
\item{object}{Celina object}

\item{cell_types_to_test}{a vector of cell types to be used for testing}

\item{scRNA_count}{a gene expression matrix of reference scRNA-seq data with gene names as rownames and cells as columns}

\item{sc_cell_type_labels}{a vector of cell type labels for each cell in \code{scRNA_count}}

\item{threshold}{a number for filtering the genes}
}
\value{
Returns Celina object.
}
\description{
Preprocess the gene expression matrix - normalize the counts and filter genes
}
