% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\name{Testing_interaction_all}
\alias{Testing_interaction_all}
\title{Functions for testing the variance term using multiple kernels}
\usage{
Testing_interaction_all(
  object,
  kernel_mat = NULL,
  method = "MM",
  celltype_to_test = NULL,
  num_cores
)
}
\arguments{
\item{object}{The SVC object.}

\item{kernel_mat}{A list of optional kernel matrix to do the testing, 
otherwise used previous calculated kernels}

\item{method}{The procedure of calculating the pvalues for each kernel, 
either by 'MM' - matching moments or by 'davies' - davies method}

\item{celltype_to_test}{Cell types selected for testing}

\item{num_cores}{The number of cores used in calculation}
}
\value{
A CELINA object with calculated testing results recorded
}
\description{
Functions for testing the variance term using multiple kernels
}
