% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELINA.R
\name{Testing_interaction_multi_kernels}
\alias{Testing_interaction_multi_kernels}
\title{Test heterogeneity term using null model for one pair. Used by Testing_interaction_all}
\usage{
Testing_interaction_multi_kernels(Y, X, covariates, kernel_mat, method)
}
\arguments{
\item{Y}{Gene expression vector}

\item{X}{celltype proportion for testing}

\item{covariates}{covariates for the null model}

\item{kernel_mat}{A list of kernel matrix to do the testing}

\item{method}{The procedure of calculating the pvalues for each kernel, 
either by 'MM' - matching moments or by 'davies' - davies method}
}
\description{
Test heterogeneity term using null model for one pair. Used by Testing_interaction_all
}
